Add files via upload
The hyothesis for the Logistic regression is the sigmoid function.   In order to optimize the parameters, instead of gradient descent steps, the inbuilt function of MATLAB fminunc is used to find the minimum of an unconstrained function.  The parameters of the cost function are being optimized.  The complete regression is evaluated using two datasets